Added Add-Column-to-Board feature (#7)
1) Added Add-Column-to-Board feature (fixed - CreateColumnCommand, BoardsController and ColumnController actions, 02-SprocsForCrud.sql and SqlStatements; 
2) Fixed EntityRepositoryBase, GetIssuesByBoardIdQuery;
3) Delete GetAllWithIssuesAndUserHandler and GetAllColumnsWithIssuesAndUserQuery.

+++ Board.BusinessLogic/Features/ForIssue/Queries/GetIssuesByBoardIdQuery.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Board.BusinessLogic.Features.ForIssue.Queries;
 
-public record GetIssuesByBoardIdQuery(int Id) : IRequest<BoardHierarchyDto>;
+public record GetIssuesByBoardIdQuery(long Id) : IRequest<BoardHierarchyDto>;
